api/requirements.txt
# --- build tools must come first ---
setuptools>=68.0.0
wheel>=0.41.0

# --- core runtime ---
flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0

# --- ML stack ---
# use a wheel that exists for Python 3.12 on Linux
numpy==1.26.4
tensorflow==2.19.0
